Refactor: moving method
