+++ lib/AppInfo/Application.php
@@ -3,22 +3,11 @@
 namespace OCA\Libresign\AppInfo;
 
 use OCA\Files\Event\LoadSidebar;
-use OCA\Libresign\Db\FileMapper;
-use OCA\Libresign\Db\FileUserMapper;
-use OCA\Libresign\Helper\JSConfigHelper;
 use OCA\Libresign\Listener\LoadSidebarListener;
-use OCA\Libresign\Storage\ClientStorage;
 use OCP\AppFramework\App;
 use OCP\AppFramework\Bootstrap\IBootContext;
 use OCP\AppFramework\Bootstrap\IBootstrap;
 use OCP\AppFramework\Bootstrap\IRegistrationContext;
-use OCP\Files\IRootFolder;
-use OCP\IL10N;
-use OCP\IRequest;
-use OCP\ISession;
-use OCP\IURLGenerator;
-use OCP\IUserManager;
-use OCP\Util;
 
 class Application extends App implements IBootstrap {
 	public const APP_ID = 'libresign';
@@ -28,7 +17,6 @@ class Application extends App implements IBootstrap {
 	}
 
 	public function boot(IBootContext $context): void {
-		$this->registerHooks($context);
 	}
 
 	public function register(IRegistrationContext $context): void {
@@ -37,35 +25,5 @@ class Application extends App implements IBootstrap {
 			LoadSidebar::class,
 			LoadSidebarListener::class
 		);
-
-		$context->registerService(ClientStorage::class, function ($c) {
-			return new ClientStorage(
-				$c->query('ServerContainer')->getUserFolder()
-			);
-		});
-	}
-	
-	private function registerHooks($context): void {
-		$request = $context->getServerContainer()->get(IRequest::class);
-		// prevent error on cron job
-		if (!$request || PHP_SAPI === 'cli' || PHP_SAPI === 'phpdbg') {
-			return;
-		}
-		$path = $request->getRawPathInfo();
-		$regex = '/' . self::APP_ID . '\/sign\/([a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12})/';
-		if (!preg_match($regex, $path)) {
-			return;
-		}
-		$jsConfigHelper = new JSConfigHelper(
-			$context->getServerContainer()->get(ISession::class),
-			$request,
-			$context->getServerContainer()->get(FileMapper::class),
-			$context->getServerContainer()->get(FileUserMapper::class),
-			$this->getContainer()->get(IL10N::class),
-			$context->getServerContainer()->get(IRootFolder::class),
-			$context->getServerContainer()->get(IURLGenerator::class),
-			$context->getServerContainer()->get(IUserManager::class)
-		);
-		Util::connectHook('\OCP\Config', 'js', $jsConfigHelper, 'extendJsConfig');
 	}
 }
